Minimal prompt theme.
Displays the $CWD only. If you are connected through ssh the hostname
will also be shown.

I like my prompt uncluttered. If I need to know if my git repo is
dirty I run `git st`. If I get lost and need to know which branch
I'm on it takes my fingers 2 seconds to type `git br<C-f>`.

The right prompt displays the last commands exit code if it was
non-zero. It requires a recent version of fish from git.