Dockerfile to install fish and test dependencies
This Dockerfile creates an ubuntu machine with fish-shell and the
dependencies we use on our test suite installed.

It does not install Oh My Fish, which one might find weird (I do). Not
confident that this is the best approach, but it's the best I could find
at the moment.