Update Easy Install instructions for lolfish theme
This commit will add a step to the "Easy Install" instructions of the `lolfish` theme.

`omf install lolfish` and `omf theme lolfish` are not enough for the `lolfish` theme to work.

Due to how `lolfish` only has a `lol.fish` file, the theme is not being recognized. Renaming this file to `fish_prompt.fish` allows for the theme to be turned on/off as usual using `omf theme <theme-name>`.

I have the command be `cp` so that the original `lol.fish` file is not lost in case other commands rely on it existing.