Fix cutting of paths with equals signs
Cut splits by equals sign and takes the second field. This ignores path
contents after a second equals sign. Currently this prevents the
installation on NixOS as the nix store path automatically contains a
"=".

This PR instead selects all but the first field and equals sign, leaving
the passed paths as intended.